cleaned up some code and files names
