+++ src/components/Controls.js
@@ -1,103 +1,109 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 
-import { createNewBoard, updateBoard } from "../functions/game";
-import { generateRandomBoard } from "../functions/presets";
+import { togglePresets } from "../functions/presets";
 
 const Controls = ({
-  board,
-  setBoard,
   gridSize,
-  isGenerating,
+  setGridSize,
+  setBoard,
   setIsGenerating,
-  genCount,
-  setGenCount,
   intervalId,
-  setInvertalId,
+  setGenCount,
   speed,
+  setSpeed,
 }) => {
-  const startGame = () => {
-    let intId = setInterval(() => {
-      let newBoard = updateBoard(board);
-      board = newBoard;
-      setBoard(newBoard);
-      setGenCount(genCount++);
-    }, speed);
-    setInvertalId(intId);
-  };
+  const [input, setInput] = useState("");
 
-  const pauseGame = () => {
+  useEffect(() => {
+    togglePresets(input, setBoard, gridSize);
+  }, [setInput, input, setBoard, gridSize]);
+
+  const handleSelect = (e) => {
+    setInput(e.target.value);
     setIsGenerating(false);
     clearInterval(intervalId);
-  };
-
-  const clearBoard = () => {
-    setIsGenerating(false);
-    setBoard(createNewBoard(gridSize));
     setGenCount(0);
-    clearInterval(intervalId);
   };
 
-  const nextGeneration = (currBoard) => {
-    let newBoard = updateBoard(currBoard);
-    board = newBoard;
-    setBoard(newBoard);
-    setGenCount((genCount += 1));
+  const handleSpeed = (e) => {
+    setSpeed(e.target.value);
+  };
+
+  const handleBoardSize = (e) => {
+    setIsGenerating(false);
+    clearInterval(intervalId);
+    setGridSize(e.target.value);
+    setGenCount(0);
   };
 
   return (
-    <div style={{ marginBottom: "10px" }}>
-      <button
-        onClick={() => {
-          if (!isGenerating) {
-            setIsGenerating(true);
-            startGame();
-          } else {
-            pauseGame();
-          }
+    <div
+      style={{
+        display: "flex",
+        alignItems: "center",
+        width: "100%",
+        justifyContent: "space-evenly",
+        marginBottom: "10px",
+      }}
+    >
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "flex-start",
         }}
-        className="gameButton"
-        style={{ background: "green" }}
       >
-        Start
-      </button>
+        <label>Presets</label>
+        <select onChange={handleSelect} value={input} className="inputField">
+          <option value="none">Choose a Preset</option>
+          <option value="acorn">Acorn</option>
+          <option value="flower of eden">Flower of Eden</option>
+          <option value="pulsar">Pulsar</option>
+          <option value="spaceship">Spaceship</option>
+          <option value="circle of fire">Circle of Fire</option>
+          <option value="glidar">Glidar</option>
+          <option value="still life">Still Life</option>
+          <option value="pulsar generator">Pulsar Generator</option>
+          <option value="r-pentomino">R-Pentomino</option>
+          <option value="pentadecatchlon">Pentadecathlon</option>
+          <option value="queen bee shuttle">Queen Bee Shuttle</option>
+        </select>
+      </div>
 
-      <button
-        onClick={pauseGame}
-        className="gameButton"
-        style={{ background: "orange" }}
-      >
-        Pause
-      </button>
-
-      <button
-        onClick={clearBoard}
-        className="gameButton"
-        style={{ background: "red" }}
-      >
-        Clear
-      </button>
-
-      <button
-        onClick={() => {
-          setGenCount(0);
-          clearInterval(intervalId);
-          setIsGenerating(false);
-          setBoard(generateRandomBoard(gridSize));
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "flex-start",
         }}
-        className="gameButton"
-        style={{ background: "purple" }}
       >
-        Random
-      </button>
-      <button
-        onClick={() => {
-          nextGeneration(board);
+        <label>Speed</label>
+        <input
+          className="inputField"
+          onChange={handleSpeed}
+          value={speed}
+          type="number"
+        />
+      </div>
+
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "flex-start",
         }}
-        className="gameButton"
-        style={{ background: "black" }}
       >
-        Next Generation
-      </button>
+        <label>Board Size</label>
+        <select
+          onChange={handleBoardSize}
+          value={gridSize}
+          className="inputField"
+        >
+          <option value={25}>Normal</option>
+          <option value={35}>Larger</option>
+          <option value={40}>Largerest</option>
+        </select>
+      </div>
     </div>
   );
 };
