+++ src/components/Header.js
@@ -4,7 +4,7 @@ const Header = () => {
   return (
     <div className="header">
       <h1>Conway's Game of Life</h1>
-      <p style={{ marginRight: "5%" }}>Aboout</p>
+      <p style={{ marginRight: "5%" }}>About</p>
     </div>
   );
 };
